Fix loop detector, NRAM behaviour... Also NRAM appearance, adding a write jumper and ensuring the write LED is only enabled when the memory is written.
The write jumper is off by default on new modules, but is set on when loaded without a defined setting (to allow old designs to work as before). The jumper setting is copyable and is saved with the file.

Additionally, switches now light up green when connected to a clock input.