Split/merge is broken again
This was my best shot at making it work intuitively, but it's not worth
the hassle. I'm switching it to a system where the direction of the
links made by the user takes precedence.

Also removed some code duplication in the operation stack - new
operations are now suppressed during undo/redo so you can use the normal
.delete() etc commands without worrying.