Consolidated most of the UI handling code
The static 'Ops' class now handles the majority of the GUI input events,
with convenient command enumerations that generate consistent UI
elements for the different menus.

It would probably be better to split up the monolithic ActionListener to
give fine-grained enable/disable, but this is much better than the old
setup.