Keep the restored window on-screen
Tested with 2 monitors in extended desktop configuration. The window is
considered "on-screen" if any part of the window intersects any of the
active display areas, which seems to be the standard way of dealing with
this. If the window is off-screen it gets restored to its default size
and centered on the active display (unless it was maximised, in which
case it remains maximised on the visible display)