Basic persistence-of-vision effect, better zoom
Dynamic refresh is capped at 30Hz, changes which occur faster than that
are shown using persistence-of-vision. In the new system, LEDs which
were turned on at any point between the last rendered refresh and this
one are displayed as being on. I also added backend support for
adjustable dynamic (simulation-driven) refresh rate, although this may
not be that useful.

Overhauled zoom to support a wider range of zoom levels and smooth zoom
in/out e.g. with trackpad. Finally fixed zoom-to-point: zooming in and
out rapidly no longer causes the camera location to jump around
randomly. Grid display is now disabled when extremely zoomed-out, which
greatly improves performance (lots of lines = bad).